# lib/api-zod/src/generated/types/speakRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SpeakRequest {
  /** Text to convert to speech */
  text: string;
  /** Language code for pronunciation guidance (e.g. en, hi, es) */
  lang: string;
}