# lib/api-zod/src/generated/types/speakResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SpeakResponse {
  /** Base64-encoded MP3 audio */
  audioBase64: string;
  /** MIME type of the audio (audio/mpeg) */
  mimeType: string;
}