# lib/api-zod/src/generated/types/transcribeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TranscribeRequest {
  /** Base64-encoded audio data (WebM or WAV) */
  audioBase64: string;
  /** MIME type of the audio (e.g. audio/webm) */
  mimeType: string;
  /** BCP-47 language code hint (e.g. en, hi, mr, es, ja, de) */
  language?: string;
}