# lib/api-zod/src/generated/types/translateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TranslateRequest {
  /** Text to translate */
  text: string;
  /** Source language code (e.g. hi, en, mr) */
  fromLang: string;
  /** Target language code (e.g. en, hi, es) */
  toLang: string;
}